chore(runtime): split regex.rs exec-array decoration under the 2000-line CI cap
#4930 (regex d flag) pushed regex.rs to 2260 lines, tripping the lint
job's file-size gate on every PR merged after it. Extract the match-
result array decoration group (index/input/groups/indices builders for
both the regex-crate fast path and the fancy_regex fallback, plus the
char/byte index converters) into regex/exec_array.rs — same recipe as
the date.rs split (#4925). No behavior change; d-flag output verified
byte-identical with Node post-split.
